pyproject.toml
[tool.poetry]
name = "webtronics"
version = "0.1.0"
description = ""
authors = ["Konstantin <kostya.kobylin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.23.1"}
SQLAlchemy = "^2.0.19"
alembic = "^1.11.1"
psycopg2 = "^2.9.6"
PyJWT = "^2.8.0"
python-decouple = "^3.8"
passlib = "^1.7.4"
SQLAlchemy-Utils = "^0.41.1"
pydantic = {extras = ["email"], version = "^2.1.1"}
bcrypt = "^4.0.1"
aiohttp = "^3.8.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
